+++ src/ResourceManager/Network/ChangeLog.md
@@ -23,6 +23,12 @@
 * Bump up Network SDK version from 18.0.0-preview to 19.0.0-preview
 
 ## Version 6.0.0
+* Added cmdlet to add a new circuit connection to an existing express route circuit.
+    - Add-AzureRmExpressRouteCircuitConnectionConfig
+* Added cmdlet to remove a circuit connection from an existing express route circuit.
+    - Remove-AzureRmExpressRouteCircuitConnectionConfig
+* Added cmdlet to retrieve a circuit connection
+    - Get-AzureRmExpressRouteCircuitConnectionConfig
 * Bump up network sdk version from 17.0.0.preview to 18.0.0.preview
 * Rerecord all network tests with latest sdk version* Set minimum dependency of module to PowerShell 5.0
 * Add support for DDoS protection plan resource